change_bot_file_structure; add_bot_thread; isort
