+++ learninglogs/learning_bot/bot/callback.py
@@ -1,0 +1,13 @@
+from telebot import types
+from telebot.callback_data import CallbackData, CallbackDataFilter
+from telebot.custom_filters import AdvancedCustomFilter
+
+groups_factory = CallbackData('group_id', prefix='groups')
+logs_factory = CallbackData('log_id', prefix='logs')
+
+
+class GroupsCallbackFilter(AdvancedCustomFilter):
+    key = 'config'
+
+    def check(self, call: types.CallbackQuery, config: CallbackDataFilter):
+        return config.check(query=call)
