+++ learninglogs/learning_bot/urls.py
@@ -1,6 +1,8 @@
 from django.urls import path
-from . import views
 from django.views.decorators.csrf import csrf_exempt
+
+from . import views
+
 app_name = 'learning_bot'
