+++ learninglogs/learninglogs/urls.py
@@ -1,8 +1,7 @@
-from django.contrib import admin
-from django.urls import include, path
 from django.conf import settings
 from django.conf.urls.static import static
-
+from django.contrib import admin
+from django.urls import include, path
 
 urlpatterns = [
     path('', include('logs.urls')),
