+++ learninglogs/users/forms.py
@@ -1,6 +1,5 @@
-from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
 from django.contrib.auth import get_user_model
-
+from django.contrib.auth.forms import AuthenticationForm, UserCreationForm
 
 User = get_user_model()
