+++ learninglogs/users/urls.py
@@ -6,6 +6,7 @@ from django.contrib.auth.views import (LoginView, LogoutView,
                                        PasswordResetDoneView,
                                        PasswordResetView)
 from django.urls import path
+
 from . import views
 
 app_name = 'users'
